Cleanup pass on the phase 3 middleware
Findings from a reuse/simplification/efficiency/altitude review of the
phase 3 diff:

- smithy_middleware.h no longer includes meerkat: MeerkatMetricsSink
  moves into the .cc behind MakeMeerkatMetricsSink() with a forward
  declaration, so only the production wiring compiles against meerkat
  (the test includes only the HttpMetricsSink seam). The library also
  drops its public visibility — package-private until a second
  smithy-cpp service needs meerkat parity.
- futility/env: new ReadPort() helper. The PORT-env read was on its
  fourth copy (meerkat, lakitu, the phase 0 greeter, and this binary),
  and portrait_smithy's copy would have been permanent.
- meerkat: metrics_manager_test now depends on :metrics_manager
  directly, completing the target split from earlier in this branch.
- portrait_smithy binary: drop the redundant runtime:core dep (arrives
  via http_beast/server), add the futility/env dep.

Reviewed and deliberately unchanged: the trace-id/access-log meerkat
duplication (documented parity; meerkat retires in phase 4), the dual
duration measurement (Observe doesn't expose headers to inner
middleware; a deeper fix belongs upstream), and the bundled
observability composite (fine with one consumer; split when hoisted).
